adding test fixture support for couchdb:
use

    couchdb_fixtures :<database_name>

to pull in the fixture file (regenerating the views may still be needed in the test class set up).

convert symbol to string. now we can use

  use_database :<database_name>

in the couchdb models.

removed blurbs in fixtures file

+++ lib/couch_rest_rails/tests.rb
@@ -2,17 +2,52 @@ module CouchRestRails
   module Tests
 
     extend self
+    mattr_accessor :fixtures_loaded
+    self.fixtures_loaded = Set.new
 
     def setup(database="*")
       ENV['RAILS_ENV'] = CouchRestRails.test_environment
-      CouchRestRails::Database.delete(database)
-      CouchRestRails::Database.create(database)
-      CouchRestRails::Fixtures.load(database)
+      unless fixtures_loaded.include?(database)
+        CouchRestRails::Database.delete(database)
+        CouchRestRails::Database.create(database)
+        CouchRestRails::Fixtures.load(database)
+        fixtures_loaded << database
+      end
+    end
+
+    def reset_fixtures
+      CouchRestRails::Database.delete("*") unless fixtures_loaded.empty?
+      fixtures_loaded.clear
     end
 
     def teardown(database="*")
       ENV['RAILS_ENV'] = CouchRestRails.test_environment
       CouchRestRails::Database.delete(database)
+      CouchRestRails::Database.create(database)
+      fixtures_loaded.delete(database)
+    end
+  end
+end
+module Test
+  module Unit #:nodoc:
+    class TestCase #:nodoc:
+      setup :setup_couchdb_fixtures
+      teardown :teardown_couchdb_fixtures
+
+      superclass_delegating_accessor :database
+      self.database = nil
+
+      class << self
+        def couchdb_fixtures(*databases)
+          self.database = databases.map { |d| d.to_s }
+        end
+      end
+      def setup_couchdb_fixtures
+        CouchRestRails::Tests.setup(self.database) unless self.database.nil?
+      end
+      def teardown_couchdb_fixtures
+        CouchRestRails::Tests.teardown(self.database) unless self.database.nil?
+      end
     end
   end
 end
